+++ Advent_Of_Code_11-20/IHeuristic.cs
@@ -3,6 +3,6 @@ namespace Advent_Of_Code_11_20
 {
     internal interface IHeuristic<TState>
     {
-        int Heuristic_Distance(Node<TState> node, Problem<TState> p);
+        float Heuristic_Distance(Node<TState> node, Problem<TState> p);
     }
 }
